refactor(capture,activities): the audience hooks live in their own files
